Include proper LaTeX class for documentation
Up until now the class file for the lyluatex package
had been used for compiling the docs, which is a bad idea
because on most computers this wouldn't be present.

This commit copies the class file(s), moves the content from
lyluatexbase.cls into a single class file - and comments out
everything that is not needed for the structure.md file.

lilypond.md may need more of lyluatex, but only after this
has been investigated unused includes or commands will finally
removed.